+++ src/elements/hero-with-background-image/src/stories.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import * as React from 'react'
-import { jsx, Styled } from 'theme-ui'
+import { jsx, Themed } from 'theme-ui'
 
 import AllThemes from '../../../utils/all-themes'
 import { Col, Row } from '../../../layout/flex-grid/src'
@@ -32,17 +32,17 @@ const content = (
   <React.Fragment>
     <Row>
       <Col span={[12, 6]}>
-        <Styled.h1 sx={{ marginTop: 0 }}></Styled.h1>
+        <Themed.h1 sx={{ marginTop: 0 }}></Themed.h1>
         <Button variant="primary">Compare investment ISAs</Button>
       </Col>
     </Row>
-    <Styled.h2
+    <Themed.h2
       sx={{
         color: 'white'
       }}
     >
       Compare and find the deal that&apos;s right for you
-    </Styled.h2>
+    </Themed.h2>
     <Row wrap direction="row">
       <Col span={[6, 3]}>
         <IconPanelHelper icon="creditCards">Credit cards</IconPanelHelper>
@@ -81,9 +81,9 @@ export const BBDealsHomepage = () => {
     <HeroWithBackgroundImage imageUrl={images.bbDeals}>
       <Row>
         <Col span={12}>
-          <Styled.h1>
+          <Themed.h1>
             Find the right broadband for you using real reviews
-          </Styled.h1>
+          </Themed.h1>
         </Col>
       </Row>
     </HeroWithBackgroundImage>