+++ src/elements/imgix-image/src/index.tsx
@@ -2,7 +2,7 @@
 
 import * as React from 'react'
 import Imgix, { SharedImigixAndSourceProps } from 'react-imgix'
-import { css, jsx } from '@emotion/core'
+import { css, jsx } from '@emotion/react'
 
 interface Props extends SharedImigixAndSourceProps {
   alt?: string