+++ src/elements/primary-info-block/src/stories.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 
-import { css, jsx } from '@emotion/core'
+import { css, jsx } from '@emotion/react'
 import { text } from '@storybook/addon-knobs'
 
 import AllThemes from '../../../utils/all-themes'