+++ src/compounds/sponsored-product/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.39](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.38...@uswitch/trustyle.sponsored-product@3.5.39) (2021-08-27)
+
+
+### Bug Fixes
+
+* fixed a11y, removed unnecessary rules, fixed package.json across all dir ([a9e0455](https://github.com/uswitch/trustyle/commit/a9e0455))
+* fixed sponsored-product container ([2d728a2](https://github.com/uswitch/trustyle/commit/2d728a2))
+* reveted changes on some packages, fixed some small bugs on themes ([69b8a3d](https://github.com/uswitch/trustyle/commit/69b8a3d))
+
+
+
+
+
 ## [3.5.38](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.37...@uswitch/trustyle.sponsored-product@3.5.38) (2021-08-19)
 
 **Note:** Version bump only for package @uswitch/trustyle.sponsored-product