+++ src/elements/accordion/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.14](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.accordion@0.14.13...@uswitch/trustyle.accordion@0.14.14) (2021-08-27)
+
+
+### Bug Fixes
+
+* drone fixed, removed theme-ui/types ([77f9479](https://github.com/uswitch/trustyle/commit/77f9479))
+* fixed a11y, removed unnecessary rules, fixed package.json across all dir ([a9e0455](https://github.com/uswitch/trustyle/commit/a9e0455))
+* reveted changes on some packages, fixed some small bugs on themes ([69b8a3d](https://github.com/uswitch/trustyle/commit/69b8a3d))
+
+
+
+
+
 ## [0.14.13](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.accordion@0.14.12...@uswitch/trustyle.accordion@0.14.13) (2021-08-12)
 
 **Note:** Version bump only for package @uswitch/trustyle.accordion