+++ src/elements/hero-secondary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-secondary@1.1.1...@uswitch/trustyle.hero-secondary@1.1.2) (2021-08-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.hero-secondary
+
+
+
+
+
 ## [1.1.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero-secondary@1.1.0...@uswitch/trustyle.hero-secondary@1.1.1) (2021-04-27)
 
 **Note:** Version bump only for package @uswitch/trustyle.hero-secondary