+++ src/styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.8](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.styles@0.5.7...@uswitch/trustyle.styles@0.5.8) (2021-08-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.styles
+
+
+
+
+
 ## [0.5.7](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.styles@0.5.6...@uswitch/trustyle.styles@0.5.7) (2021-04-27)
 
 **Note:** Version bump only for package @uswitch/trustyle.styles